# resources/openapi/base/environment.yaml
openapi: 3.1.0
info:
  title: Cbox ID — Environment Management API
  version: "1.0.0"
  description: |
    The **per-environment management plane** — the API an app's own backend uses to run
    its tenancy inside ONE environment: organizations and their owners, members,
    invitations, roles (in one organization, or everywhere for your staff), apps, APIs,
    the API keys customers made for them, and support sessions. It is served on
    the environment's own host ({slug}.cboxid.com or a custom domain): the host pins
    the environment, and the request is authenticated with an environment API key
    (`cbid_env_…`).

    This is distinct from the global organization API (`cbid_ws_…`), which manages the
    organization, its environments and members. Credentials never cross planes — an
    organization key cannot call this API, and an environment key cannot call the organization
    API. An environment key is bound to the host it was minted for: presenting it on
    another environment's host fails authentication outright.

    **Scopes.** A key carries an explicit allow-list; every endpoint requires one
    (deny-by-default). `resource:read` never implies `resource:write`, so a
    read-only integration key cannot mutate.

    **Envelope.** Every success returns `{ "data": … }` (an object, or an array with
    a sibling `meta` on lists). Lists are cursor-paginated: pass `?after=<next_cursor>`
    from the previous page's `meta`. Every error returns
...

...

...

...

...

...

...
servers:
...
security:
  - EnvironmentApiKey: []
...
paths:
...
  /apps:
    get:
      tags: [Apps]
      summary: List apps
      description: "Requires scope `apps:read`. Never a secret."
      operationId: apps_list
      x-action: apps.list
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/After"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/App" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
    post:
      tags: [Apps]
      summary: Register an app
      description: |
        Requires scope `apps:write`. Two ways to describe it:

        - **From a blueprint** — send `GET /apps/{id}/blueprint`'s document from another
          environment as `blueprint` (promote staging to production). `name`,
          `redirect_uris` and `post_logout_redirect_uris` beside it replace the
          blueprint's own, since those usually name the environment it came from. A
          `private_key_jwt` blueprint needs `jwks`.
        - **Short form** — `name` and `type` (`web`, `spa`, `cli`, `service`, `agent`; default
          `web`), which decide the client type, grants and default scopes as the console's
          "Create app" does. `advanced` takes `client_type` and `grant_types`.

        The response carries `client_secret` for an app that authenticates with one —
        **once**. Only its hash is kept, and an idempotent replay returns `client_secret: null`.

        A platform scope reserved for the console (`vault.manage`, `decisions:read`) is
        refused (`scope_not_grantable`), as is a registered API's scope this app's owner may
        not hold. Danger: critical — it mints a credential.
      operationId: apps_create
      x-action: apps.create
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                blueprint: { $ref: "#/components/schemas/AppBlueprint" }
                name: { type: string, maxLength: 190 }
                type: { type: string, enum: [web, spa, cli, service, agent, advanced], default: web }
                client_type: { type: string, enum: [confidential, public] }
                grant_types: { type: array, items: { type: string } }
                redirect_uris: { type: array, items: { type: string } }
                post_logout_redirect_uris: { type: array, items: { type: string } }
                scopes: { type: array, items: { type: string } }
                first_party: { type: boolean, default: false }
                manifest_url: { type: ["string", "null"], format: uri, maxLength: 500, description: "Where the app publishes its roles-and-permissions manifest (short form)." }
                organization_id: { type: ["string", "null"], description: "The owning organization; null for the environment's own app." }
                jwks: { type: ["object", "null"], description: "A public JWK Set, for a `private_key_jwt` app." }
      responses:
        "201":
          description: Registered.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/App" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /apis:
    get:
      tags: [APIs]
      summary: List registered APIs
      description: "Requires scope `apis:read`."
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/After"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Api" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
    post:
      tags: [APIs]
      summary: Register an API and its scopes
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      description: |
        Requires scope `apis:write`. Registering an API is the environment's act; no tenant
        surface can. `identifier` is an absolute URI (RFC 8707) and becomes the token's
        `aud`. Scope keys are unique across the environment. `client_id` links the app whose
        roles and permissions the API enforces, and must have the API's owner.
        `organization_id` makes the API one organization's; left out, the environment owns
        it. Refusals are `422 invalid_api` with the reason.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, name]
              properties:
                identifier: { type: string, format: uri, maxLength: 255 }
                name: { type: string, maxLength: 120 }
                organization_id: { type: ["string", "null"] }
                client_id: { type: ["string", "null"] }
                scopes:
                  type: array
                  items: { $ref: "#/components/schemas/ApiScope" }
      responses:
        "201":
          description: Registered.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/Api" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /apis/{id}:
    get:
      tags: [APIs]
      summary: Get an API
      description: "Requires scope `apis:read`."
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/Api" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
    patch:
      tags: [APIs]
      summary: Change an API
      description: |
        Requires scope `apis:write`. Any of: `name`; `client_id` (null unlinks the app);
        `scopes`, which is the **complete** set afterwards — scopes named are added or
        updated, scopes left out are removed. All or nothing. The identifier never changes:
        register a new API instead.
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 120 }
                client_id: { type: ["string", "null"] }
                scopes:
                  type: array
                  items: { $ref: "#/components/schemas/ApiScope" }
      responses:
        "200":
          description: The API after the change.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/Api" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
    delete:
      tags: [APIs]
      summary: Delete an API and its scopes
      description: |
        Requires scope `apis:write`. Tokens already minted for it keep their `aud` until they
        expire; apps holding its scope keys keep them as plain scopes.
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204": { description: "Deleted." }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /action-approvals/{id}:
    get:
      tags: [Approvals]
      summary: Where an approval this key asked for stands
      description: |
        Any key may poll the approvals it raised, and only those: another key's id is a 404.
        `status` is `pending`, `approved`, `denied`, `expired` or `consumed`.
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [id, status]
                    properties:
                      id: { type: string }
                      status: { type: string, enum: [pending, approved, denied, expired, consumed] }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /vault/secrets:
    post:
      tags: [Token Vault]
      summary: Store a downstream credential, sealed at rest
      description: |
        The secret is sealed with the environment's crypto key and is never returned by
        any read endpoint — only a `lease` reveals it, and only to a client holding a
        live grant.

        Ownership is derived from the calling token's `org` claim and CANNOT be supplied
        in the body: the owner is the vault's tenancy boundary, and a caller-chosen owner
        would be no boundary at all.
      security: [{ VaultManagerToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, provider, secret]
              properties:
                name:
                  type: string
                  maxLength: 200
                  description: Unique per owner — two organizations may each hold a `github-token`.
                provider: { type: string, maxLength: 100 }
                secret: { type: string, description: The value to seal. Never returned by a read. }
                expires_at: { type: ["string", "null"], format: date-time }
      responses:
        '201':
          description: Stored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VaultSecret' }
        '401': { $ref: '#/components/responses/BearerUnauthorized' }
        '403': { $ref: '#/components/responses/BearerForbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /vault/secrets/{id}/rotate:
    post:
      tags: [Token Vault]
      summary: Replace the sealed value, keeping the id and its grants
      security: [{ VaultManagerToken: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [secret]
              properties:
                secret: { type: string }
      responses:
        '200':
          description: Rotated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VaultSecret' }
        '401': { $ref: '#/components/responses/BearerUnauthorized' }
        '403': { $ref: '#/components/responses/BearerForbidden' }
        '404':
          description: |
            No such secret **in the caller's organization**. A secret owned by another
            organization is indistinguishable from one that does not exist — the lookup
            is filtered, not fetched-then-compared, so nothing leaks about what exists
            outside the caller's own scope. The message is a constant for the same reason.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: not_found, message: "No such secret." }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /vault/secrets/{id}:
    delete:
      tags: [Token Vault]
      summary: Revoke a secret permanently
      description: No future lease can open it. Existing leases are not recalled — they are short-lived by design.
      security: [{ VaultManagerToken: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: Revoked. }
        '401': { $ref: '#/components/responses/BearerUnauthorized' }
        '403': { $ref: '#/components/responses/BearerForbidden' }
        '404':
          description: No such secret in the caller's organization.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /vault/secrets/{id}/grants:
    post:
      tags: [Token Vault]
      summary: Authorize an agent client to lease this secret
      security: [{ VaultManagerToken: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id]
              properties:
                client_id: { type: string, maxLength: 200 }
                max_ttl_seconds: { type: ["integer", "null"], minimum: 1 }
      responses:
        '201':
          description: Granted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret_id: { type: string }
                  client_id: { type: string }
                  max_ttl_seconds: { type: ["integer", "null"] }
        '401': { $ref: '#/components/responses/BearerUnauthorized' }
        '403': { $ref: '#/components/responses/BearerForbidden' }
        '404':
          description: No such secret in the caller's organization.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /vault/secrets/{id}/grants/{clientId}:
    delete:
      tags: [Token Vault]
      summary: Revoke an agent's authorization
      description: Idempotent — revoking an absent grant succeeds.
      security: [{ VaultManagerToken: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
        - name: clientId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Revoked (or already absent). }
        '401': { $ref: '#/components/responses/BearerUnauthorized' }
        '403': { $ref: '#/components/responses/BearerForbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /vault/secrets/{id}/lease:
    post:
      tags: [Token Vault]
      summary: Redeem a grant for the plaintext secret
      description: |
        The only endpoint that reveals a secret. Every refusal — unknown secret, revoked,
        expired, or no live grant — is the SAME 403 with no detail, so a caller cannot
        use it to enumerate what exists.
      security: [{ AgentToken: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [purpose]
              properties:
                purpose: { type: string, maxLength: 200, description: Recorded on the audit trail. }
      responses:
        '200':
          description: Leased.
          content:
            application/json:
              schema:
                type: object
                required: [secret_id, provider, secret, expires_at]
                properties:
                  secret_id: { type: string }
                  provider: { type: string, description: The downstream provider the secret belongs to. }
                  secret: { type: string }
                  expires_at: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/BearerUnauthorized' }
        '403':
          description: |
            Lease denied. Deliberately uniform — unknown secret, revoked, expired and
            ungranted are ONE response, with a constant message, so it cannot be used to
            enumerate. A token missing the `vault.lease` scope is an RFC 6750 challenge
            instead, and is distinguishable only because it never reaches the vault.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Error'
                  - $ref: '#/components/schemas/BearerChallenge'
              example: { error: lease_denied, message: "The lease was denied." }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    EnvironmentApiKey:
      type: http
      scheme: bearer
      bearerFormat: cbid_env_
      description: |
        An environment API key, `cbid_env_…`. Created by the account that owns the environment,
        in the console under Identity platform › Environment keys — not in the environment's own
        console. Bound to that environment's host, and carries an explicit list of scopes; it may
        also carry an expiry.
    ManagementAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A PERSON's OAuth access token, issued by this environment's issuer and audienced to
        its management plane (`resource=https://<environment-host>/mcp` — the same token the
        MCP server takes). `cbox login` and MCP clients obtain one by signing the person in.

        It may run an operation only when it carries the operation's scope AND the person
        could run it on their own console here: an organization administrator, confined to
        the organization the token is bound to. Operations that belong to the environment
        console do not list it. Every critical operation answers `202 approval_required`
        until the person approves it on their device.

        On the organization's own roster (members, their roles, invitations, ownership, its
        people's API keys) the organization's own rules hold, as on its People page: an
        admin cannot change or remove an owner, nobody removes themselves, only the owner
        transfers ownership, a staff-only role is `role_not_assignable`, and a customer
        workspace's roster — administered under Workspace › Team — is `409 managed_elsewhere`. A DPoP-bound token is presented as
        `Authorization: DPoP <token>` with a proof.
    WorkspaceAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A workspace member's OAuth access token from the PLATFORM ROOT, audienced to the
        root's `/mcp` (`cbox login` at the root; an agent signed in there). Presented on the
        ROOT's host, with `Cbox-Environment` naming the environment of the person's
        workspace to act in. It runs what that environment's console runs — when the
        person's role administers environments and their membership reaches this one — and
...
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Any unique string (a UUID). A retry with the same key and request gets the first answer back.
      schema: { type: string, maxLength: 255 }
...
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    After:
      name: after
      in: query
      required: false
      description: Opaque cursor — the `meta.next_cursor` from the previous page.
      schema: { type: string }
  schemas:
...
    App:
      type: object
      required: [id, client_id, name, type, client_type, grant_types, redirect_uris, scopes, first_party]
      properties:
        id: { type: string }
        client_id: { type: string }
        name: { type: string }
        type:
          type: string
          enum: [web, spa, cli, service, agent, advanced]
          description: The kind, read back from its grants; `advanced` when they match no preset.
        client_type: { type: string, enum: [confidential, public] }
        organization_id: { type: ["string", "null"] }
        first_party: { type: boolean }
        grant_types: { type: array, items: { type: string } }
        redirect_uris: { type: array, items: { type: string } }
        post_logout_redirect_uris: { type: array, items: { type: string } }
        scopes: { type: array, items: { type: string } }
        manifest_url: { type: ["string", "null"], description: Where it publishes its roles-and-permissions manifest. }
        access_token_ttl: { type: ["integer", "null"], description: Its access-token lifetime in seconds; null for the install's default. }
        backchannel_logout_uri: { type: ["string", "null"], description: Where it is sent a logout token (OIDC Back-Channel Logout); null when it is not told. }
        backchannel_logout_session_required: { type: boolean }
        api_key_prefix: { type: ["string", "null"], description: The prefix of its customer API keys; null when it accepts none. }
        created_at: { type: ["string", "null"], format: date-time }
        client_secret:
          type: ["string", "null"]
          description: Only on the response that created (or copied) the app, and only for an app that authenticates with a secret. Never again — null on an idempotent replay.
...
    AppBlueprint:
      type: object
      description: |
        An app's configuration without its identity or credentials — deterministic (lists
        sorted), versioned, and refused rather than half-read when it carries a key this
        server does not know.
      required: [kind, version, name, client_type]
      additionalProperties: false
      properties:
        kind: { type: string, const: cbox-id.client-blueprint }
        version: { type: integer, const: 1 }
        name: { type: string }
        client_type: { type: string, enum: [confidential, public] }
        token_endpoint_auth_method: { type: ["string", "null"] }
        grant_types: { type: array, items: { type: string } }
        redirect_uris: { type: array, items: { type: string } }
        post_logout_redirect_uris: { type: array, items: { type: string } }
        scopes: { type: array, items: { type: string } }
        first_party: { type: boolean }
        manifest_url: { type: ["string", "null"] }
        access_token_ttl: { type: ["integer", "null"] }
        backchannel_logout_uri: { type: ["string", "null"] }
        backchannel_logout_session_required: { type: boolean }
        api_key_prefix: { type: ["string", "null"] }
...
    PageMeta:
      type: object
      properties:
        limit: { type: integer }
        has_more: { type: boolean }
        next_cursor: { type: ["string", "null"] }
...
  responses:
...
    Unauthorized:
      description: Missing or invalid environment API key.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The key is missing the scope this endpoint requires.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    TooManyRequests:
      description: |
        The key's rate limit for this plane is exhausted. The bucket is keyed on the API
        KEY, so this reflects only your own traffic.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...